Ship abac in the bitstream: GP18, entropy type 5, −17.3% at q=90
The step BACKLOG "EBCOT — evaluating in halves" Part 6 left open. abac was a
standalone module and a diagnostic; it is now a selectable entropy coder that
writes and reads real files.

`EntropyCoder::Abac` / `--abac`, bitstream `entropy_type = 5` behind a new GP18
generation. GP18 adds nothing else, so a GP18 frame using any older coder is
byte-identical to the GP17 one apart from four magic bytes — but a GP17 decoder
rejects type 5 rather than misreading it, which is what the generation is for.

Rate against shipped Rice, measured through encode → file → GPU decode on four
images (bbb, blue_sky, kristensara, touchdown). Entropy coding is lossless, so
both arms decode to **identical pixels** and the rate delta is exact rather than
a rate/quality trade:

| q  | bbb    | blue_sky | kristensara | touchdown | mean   |
|----|--------|----------|-------------|-----------|--------|
| 50 | −17.8% | −18.5%   | −19.4%      | −19.5%    | −18.8% |
| 75 | −14.7% | −15.8%   | −19.3%      | −16.8%    | −16.6% |
| 90 | −14.2% | −17.2%   | −20.9%      | −16.9%    | −17.3% |
| 100| −14.2% | −14.5%   | −15.0%      | −10.0%    | −13.4% |

q=90 was predicted at −16.7% from `GNC_ABAC_COMPARE` on the same four images;
the real bitstream measures −17.3%, slightly better, consistent with abac's 25
two-byte block headers against Rice's per-tile header plus 256 length fields.

q=100 is the MED lossless path, still bit-exact. It takes the FFV1 gap from
+23.9% to **+7.3%**, against an FFV1 level-3 gbrp encode run today on the same
four images.

q=25 is deliberately not quoted as a rate figure: the two arms differ by 0.12 dB
there, and not because of abac — see below.

Correctness. Rice and abac code the same quantised coefficients, so a divergence
is visible as different pixels rather than as worse ones. bbb at q=90 decodes
pixel-identical through the container (max |diff| 0 over 1920x1080x3), and
`tests/abac_bitstream.rs` asserts that at q=50/75/90, at 4:2:2 and 4:2:0, and
across both arithmetic engines. This matters more than a usual equivalence test:
an adaptive arithmetic decoder that diverges by one bit does not fail, it reads
every later symbol from a corrupted interval *and* a corrupted context and
reconstructs a plausible wrong image.

One real bug found on the way, and it is the kind that does not raise an error:
the decode shader wrote `array<i32>` into `scratch_a`, which every other entropy
decoder writes as f32. Negative coefficients became NaN, because -1 as i32 is
0xFFFFFFFF, a quiet NaN as f32. Rate was already correct at that point; only the
picture was gone.

Not abac, found while testing it and now pinned by a test so nobody re-finds it
as an abac bug: **Rice's GPU encode path and its CPU encode path do not produce
the same pixels** — 35.51 vs 35.63 dB on bbb at q=25, and max |diff| 1.69 at
4:2:2. Both arms are Rice, so the entropy coder is not the difference; the fused
quantize+histogram shader runs only on the GPU encode path.

Canary (`GNC_DIAGNOSTICS=1`): `[abac] plane 8x5 tiles: abac_blocks=1000
(empty=0) bytes=769006 coder=Range cb=64` per plane.

Defaults: Range at cb=64, which dominates every other cell measured. Encode is
CPU (a serial adaptive coder), decode is GPU, one thread per code-block.
Decode throughput is the known ~1.69x debt and was not re-measured here.

Intra only. Inter frames reuse the same path but residual statistics differ and
the contexts are untuned for them.

+++ src/shaders/abac_decode.wgsl
@@ -52,7 +52,11 @@ struct Params {
 @group(0) @binding(0) var<uniform> params: Params;
 @group(0) @binding(1) var<storage, read> stream: array<u32>;
 @group(0) @binding(2) var<storage, read> blocks: array<BlockInfo>;
-@group(0) @binding(3) var<storage, read_write> out: array<i32>;
+// f32, not i32: this is the same buffer every other entropy decoder writes (`scratch_a`), and the
+// dequantiser downstream reads it as float. Writing raw i32 bits here decodes without error and
+// then turns every negative coefficient into a NaN, because -1 as i32 is 0xFFFFFFFF, which is a
+// quiet NaN as f32. Quantised coefficients are far inside f32's exact-integer range.
+@group(0) @binding(3) var<storage, read_write> out: array<f32>;
 
 // Decoder state, per thread. Kept in one struct so the helpers read like the Rust ones.
 struct Dec {
@@ -271,7 +275,7 @@ fn main(
                     v = i32(a);
                 }
             }
-            out[info.out_offset + y * info.stride + x] = v;
+            out[info.out_offset + y * info.stride + x] = f32(v);
             rows[(cur + x) * WG + tid] = a;
         }
     }
@@ -435,7 +439,7 @@ fn main_rc(
                     v = i32(a);
                 }
             }
-            out[info.out_offset + y * info.stride + x] = v;
+            out[info.out_offset + y * info.stride + x] = f32(v);
             rows[(cur + x) * WG + tid] = a;
         }
     }
